ignition-layout-studio/backend/sandbox/requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
pillow>=8.3.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=1.0.0
plotly>=5.0.0
seaborn>=0.11.0
openpyxl>=3.0.0
xlrd>=2.0.0